visual_groups: move VisualGroupForm to Views/, drop _form partial
Components::VisualGroupForm -> Views::Controllers::VisualGroups::Form
Per #4331. The _form.html.erb wrapper partial is deleted; new.html.erb
and edit.html.erb render the Phlex view directly.
