--- app/views/controllers/visual_groups/_form.html.erb
@@ -1,4 +1,0 @@
-<%= render(Components::VisualGroupForm.new(
-  visual_group,
-  visual_model: @visual_model
-)) %>
