+++ app/views/controllers/visual_groups/edit.html.erb
@@ -12,7 +12,9 @@ container_class(:full)
                 visual_model_visual_groups_path(@visual_group.visual_model)) %>
   </p>
 
-  <%= render('form', visual_group: @visual_group )%>
+  <%= render(Views::Controllers::VisualGroups::Form.new(
+        @visual_group, visual_model: @visual_group.visual_model
+      )) %>
 
   <span id="filter_options">
     <hr>
